refactor: Update package name across the project
This commit refactors the entire project to use the new package name `com.quare.bibleplanner`, replacing the old `quare.software.bibleplanner`.

The package name has been updated in:
*   All Kotlin source files across `composeApp`, `core`, and `ui` modules.
*   `build.gradle.kts` files for `composeApp`, `core`, and `ui` modules to update namespaces and application identifiers.
*   The iOS `Config.xcconfig` file to change the product bundle identifier.
