--- composeApp/src/androidMain/kotlin/quare/software/bibleplanner/Platform.android.kt
@@ -1,9 +1,0 @@
-package quare.software.bibleplanner
-
-import android.os.Build
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
