+++ composeApp/src/commonMain/kotlin/com/quare/bibleplanner/App.kt
@@ -1,0 +1,30 @@
+package com.quare.bibleplanner
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import org.jetbrains.compose.ui.tooling.preview.Preview
+import com.quare.bibleplanner.ui.theme.AppTheme
+
+@Composable
+@Preview
+fun App() {
+    AppTheme {
+        Surface(
+            modifier = Modifier.fillMaxSize(),
+            color = MaterialTheme.colorScheme.background,
+        ) {
+            Box(
+                modifier = Modifier.fillMaxSize(),
+                contentAlignment = Alignment.Center
+            ) {
+                Text(text = "Hello World!")
+            }
+        }
+    }
+}
