--- composeApp/src/commonMain/kotlin/quare/software/bibleplanner/Greeting.kt
@@ -1,9 +1,0 @@
-package quare.software.bibleplanner
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
