--- composeApp/src/commonMain/kotlin/quare/software/bibleplanner/Platform.kt
@@ -1,7 +1,0 @@
-package quare.software.bibleplanner
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
