+++ composeApp/src/iosMain/kotlin/com/quare/bibleplanner/MainViewController.kt
@@ -1,0 +1,5 @@
+package com.quare.bibleplanner
+
+import androidx.compose.ui.window.ComposeUIViewController
+
+fun MainViewController() = ComposeUIViewController { App() }
