--- composeApp/src/iosMain/kotlin/quare/software/bibleplanner/MainViewController.kt
@@ -1,5 +1,0 @@
-package quare.software.bibleplanner
-
-import androidx.compose.ui.window.ComposeUIViewController
-
-fun MainViewController() = ComposeUIViewController { App() }
