--- composeApp/src/iosMain/kotlin/quare/software/bibleplanner/Platform.ios.kt
@@ -1,9 +1,0 @@
-package quare.software.bibleplanner
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
