+++ composeApp/src/jvmMain/kotlin/com/quare/bibleplanner/main.kt
@@ -1,0 +1,13 @@
+package com.quare.bibleplanner
+
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.application
+
+fun main() = application {
+    Window(
+        onCloseRequest = ::exitApplication,
+        title = "BiblePlanner",
+    ) {
+        App()
+    }
+}
