--- composeApp/src/jvmMain/kotlin/quare/software/bibleplanner/Platform.jvm.kt
@@ -1,7 +1,0 @@
-package quare.software.bibleplanner
-
-class JVMPlatform: Platform {
-    override val name: String = "Java ${System.getProperty("java.version")}"
-}
-
-actual fun getPlatform(): Platform = JVMPlatform()
