--- composeApp/src/jvmMain/kotlin/quare/software/bibleplanner/main.kt
@@ -1,13 +1,0 @@
-package quare.software.bibleplanner
-
-import androidx.compose.ui.window.Window
-import androidx.compose.ui.window.application
-
-fun main() = application {
-    Window(
-        onCloseRequest = ::exitApplication,
-        title = "BiblePlanner",
-    ) {
-        App()
-    }
-}
