+++ core/model/src/commonMain/kotlin/com/quare/bibleplanner/core/model/Model.kt
@@ -1,0 +1,3 @@
+package com.quare.bibleplanner.core.model
+
+// Add your data models here
