--- core/model/src/commonMain/kotlin/quare/software/bibleplanner/core/model/Model.kt
@@ -1,3 +1,0 @@
-package quare.software.bibleplanner.core.model
-
-// Add your data models here
