+++ core/navigation/build.gradle.kts
@@ -4,15 +4,12 @@ plugins {
 }
 
 android {
-    namespace = "quare.software.bibleplanner.core.navigation"
+    namespace = "com.quare.bibleplanner.core.navigation"
 }
 
 kotlin {
     sourceSets {
         commonMain.dependencies {
-            // Add navigation library dependency here when needed
-            // e.g., Voyager, Decompose, or Compose Navigation
         }
     }
 }
-
