+++ core/navigation/src/commonMain/kotlin/com/quare/bibleplanner/core/navigation/Navigation.kt
@@ -1,0 +1,3 @@
+package com.quare.bibleplanner.core.navigation
+
+// Add your navigation configuration here
