--- core/navigation/src/commonMain/kotlin/quare/software/bibleplanner/core/navigation/Navigation.kt
@@ -1,3 +1,0 @@
-package quare.software.bibleplanner.core.navigation
-
-// Add your navigation configuration here
