+++ core/provider/koin/src/commonMain/kotlin/com/quare/bibleplanner/core/provider/koin/KoinProvider.kt
@@ -1,0 +1,3 @@
+package com.quare.bibleplanner.core.provider.koin
+
+// Add your Koin dependency injection setup here
