--- core/provider/koin/src/commonMain/kotlin/quare/software/bibleplanner/core/provider/koin/KoinProvider.kt
@@ -1,3 +1,0 @@
-package quare.software.bibleplanner.core.provider.koin
-
-// Add your Koin dependency injection setup here
