+++ iosApp/Configuration/Config.xcconfig
@@ -1,7 +1,7 @@
 TEAM_ID=
 
 PRODUCT_NAME=BiblePlanner
-PRODUCT_BUNDLE_IDENTIFIER=quare.software.bibleplanner.BiblePlanner$(TEAM_ID)
+PRODUCT_BUNDLE_IDENTIFIER=com.quare.bibleplanner.BiblePlanner$(TEAM_ID)
 
 CURRENT_PROJECT_VERSION=1
 MARKETING_VERSION=1.0
