+++ ui/theme/src/commonMain/kotlin/com/quare/bibleplanner/ui/theme/ThemeOption.kt
@@ -1,0 +1,13 @@
+package com.quare.bibleplanner.ui.theme
+
+import androidx.compose.runtime.compositionLocalOf
+
+val LocalThemeOption = compositionLocalOf {
+    ThemeOption.SYSTEM
+}
+
+enum class ThemeOption {
+    LIGHT,
+    DARK,
+    SYSTEM,
+}
