+++ ui/theme/src/commonMain/kotlin/com/quare/bibleplanner/ui/theme/color/ColorScheme.kt
@@ -1,0 +1,7 @@
+package com.quare.bibleplanner.ui.theme.color
+
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+
+internal val lightScheme = lightColorScheme()
+internal val darkScheme = darkColorScheme()
