--- ui/theme/src/commonMain/kotlin/quare/software/bibleplanner/ui/theme/Theme.kt
@@ -1,26 +1,0 @@
-package quare.software.bibleplanner.ui.theme
-
-import androidx.compose.material3.ColorScheme
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import quare.software.bibleplanner.ui.theme.color.darkScheme
-import quare.software.bibleplanner.ui.theme.color.lightScheme
-
-@Composable
-fun AppTheme(
-    getSpecificColors: @Composable ((Boolean) -> ColorScheme?)? = null,
-    content: @Composable () -> Unit
-) {
-    val isDarkTheme = isAppInDarkTheme()
-    MaterialTheme(
-        colorScheme = getSpecificColors?.invoke(isDarkTheme) ?: getColorScheme(isDarkTheme),
-        content = content
-    )
-}
-
-@Composable
-private fun getColorScheme(isDarkTheme: Boolean): ColorScheme = if (isDarkTheme) {
-    darkScheme
-} else {
-    lightScheme
-}
