--- ui/theme/src/commonMain/kotlin/quare/software/bibleplanner/ui/theme/ThemeExt.kt
@@ -1,11 +1,0 @@
-package quare.software.bibleplanner.ui.theme
-
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.runtime.Composable
-
-@Composable
-fun isAppInDarkTheme(): Boolean = when (LocalThemeOption.current) {
-    ThemeOption.LIGHT -> false
-    ThemeOption.DARK -> true
-    ThemeOption.SYSTEM -> isSystemInDarkTheme()
-}
