--- ui/theme/src/commonMain/kotlin/quare/software/bibleplanner/ui/theme/ThemeOption.kt
@@ -1,13 +1,0 @@
-package quare.software.bibleplanner.ui.theme
-
-import androidx.compose.runtime.compositionLocalOf
-
-val LocalThemeOption = compositionLocalOf {
-    ThemeOption.SYSTEM
-}
-
-enum class ThemeOption {
-    LIGHT,
-    DARK,
-    SYSTEM,
-}
