--- ui/theme/src/commonMain/kotlin/quare/software/bibleplanner/ui/theme/color/ColorScheme.kt
@@ -1,7 +1,0 @@
-package quare.software.bibleplanner.ui.theme.color
-
-import androidx.compose.material3.darkColorScheme
-import androidx.compose.material3.lightColorScheme
-
-internal val lightScheme = lightColorScheme()
-internal val darkScheme = darkColorScheme()
